Move accept conn logic to downstream + upstream
